Remove Grid.isPartialLastRow().
The semantics were not clear in relation to hidden rows, so we make rows have
"visibility: hidden" instead of "display: none" so that their height is known,
and add (hidden) rows until more clearly-defined conditions are met, so that
we know how far we can scroll the grid and still display all rows on screen.

Update setScroll() and addRow() logic accordingly.